fix done

+++ routes/admin/penilaian.php
@@ -7,11 +7,11 @@ Route::group(['middleware' => ['auth']], function () {
 
     Route::controller(PenilaianController::class)->group(function(){
         Route::get('penilaian', [PenilaianController::class, 'index'])->name('penilaian');
-        Route::get('penilaian/tambah', [PenilaianController::class, 'create'])->name('penilaian.tambah');
+        Route::get('penilaian/tambah', [PenilaianController::class, 'create'])->name('penilaian.tambah')->middleware(['auth','role.custom:pengunjung']);
         Route::get('penilaian/detail/{id}', [PenilaianController::class, 'show'])->name('penilaian.detail');
-        Route::delete('penilaian/{id}', [PenilaianController::class, 'destroy'])->name('penilaian.hapus');
-        Route::post('penilaian/store', [PenilaianController::class, 'store'])->name('penilaian.store');
-        Route::get('penilaian/{id}/ubah', [PenilaianController::class, 'edit'])->name('penilaian.ubah');
-        Route::put('penilaian/{id}', [PenilaianController::class, 'update'])->name('penilaian.update');
+        Route::delete('penilaian/{id}', [PenilaianController::class, 'destroy'])->name('penilaian.hapus')->middleware(['auth','role.custom:pengunjung|admin']);
+        Route::post('penilaian/store', [PenilaianController::class, 'store'])->name('penilaian.store')->middleware(['auth','role.custom:pengunjung']);
+        Route::get('penilaian/{id}/ubah', [PenilaianController::class, 'edit'])->name('penilaian.ubah')->middleware(['auth','role.custom:pengunjung']);
+        Route::put('penilaian/{id}', [PenilaianController::class, 'update'])->name('penilaian.update')->middleware(['auth','role.custom:pengunjung']);
     });
 });
